Deleted .gitKeep
